🩹 [Patch]: Add nightly run, git settings and linting settings (#4)
## Description

This pull request includes several changes to improve the configuration
and functionality of the repository. The most important changes include
updates to `.gitattributes`, new linter configuration files,
modifications to GitHub Actions workflows, and changes to the
`scripts/main.ps1` file.

### Configuration updates:

*
[`.gitattributes`](diffhunk://#diff-618cd5b83d62060ba3d027e314a21ceaf75d36067ff820db126642944145393eL1-R15):
Updated to include specific end-of-line configurations for different
file types and ensure binary files are treated correctly.

### Linter configuration:

*
[`.github/linters/.jscpd.json`](diffhunk://#diff-557094e283c00b23265c1c75872f41c6b1a524a00f0d99dd68ebd22cb63bfdd6R1-R10):
Added a new configuration file for the JavaScript Copy/Paste Detector
(jscpd) with specified thresholds and ignore patterns.
*
[`.github/linters/.markdown-lint.yml`](diffhunk://#diff-2fd3619d2b7372123257bc21645edb90c24293e66babbc6053033a89e21bab34R1-R25):
Added a new configuration file for the Markdown linter with specified
rules and exceptions.
*
[`.github/linters/.powershell-psscriptanalyzer.psd1`](diffhunk://#diff-aae69c9d6774628ed181eacf53aee0f38eb6c2f53492cf3a5b7f7bdb6ef43b6aR1-R56):
Added a new configuration file for the PowerShell Script Analyzer with
specified rules and exclusions.

### GitHub Actions workflows:

*
[`.github/workflows/Action-Test.yml`](diffhunk://#diff-a12ae5c885b0673c0ff6f70c2670886907590d624626e07da4c52e01aeaf56a4L5-R9):
Modified to include `workflow_dispatch` and `schedule` triggers, and
updated secret references for `ClientID` and `PrivateKey`.
[[1]](diffhunk://#diff-a12ae5c885b0673c0ff6f70c2670886907590d624626e07da4c52e01aeaf56a4L5-R9)
[[2]](diffhunk://#diff-a12ae5c885b0673c0ff6f70c2670886907590d624626e07da4c52e01aeaf56a4L28-R33)
*
[`.github/workflows/Linter.yml`](diffhunk://#diff-482e65806ed9e4a7320f14964764086b91fed4a28d12e4efde1776472e147e79L27-R30):
Updated to use the latest version of `super-linter` and added
environment variables to disable JSON, Markdown, and YAML Prettier
validation.

### Script updates:

*
[`scripts/main.ps1`](diffhunk://#diff-dc2e5a659836b1b73abb03421c567f5018c2755677c4a0aa764cb26117b68011L2-R2):
Refactored to remove parameter declarations and directly use environment
variables for `ClientID` and `PrivateKey`. Improved output handling for
JWT generation.
[[1]](diffhunk://#diff-dc2e5a659836b1b73abb03421c567f5018c2755677c4a0aa764cb26117b68011L2-R2)
[[2]](diffhunk://#diff-dc2e5a659836b1b73abb03421c567f5018c2755677c4a0aa764cb26117b68011L27-R28)
[[3]](diffhunk://#diff-dc2e5a659836b1b73abb03421c567f5018c2755677c4a0aa764cb26117b68011L45-R39)

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas